Normalize url config value
The `Url` trait treats trailing slashes as significant (https://docs.rs/url/*/url/struct.Url.html#method.join).

This normalizes the url setting when using via the API so users don't have
to always enter trailing slashes to get proper behavior.

We could have done this when loading the config but we may need to retain
the original user-entered value for some reason so instead we are normalizing
at the API interaction layer.